requirements.txt
anthropic>=0.40.0
python-telegram-bot>=20.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
aiohttp>=3.9.0
